Move over_react component into lib
+ The transformer does not run in web/ by default, which is what caused the issue reported via https://github.com/Workiva/over_react/issues/148
